p058: fixed

+++ 058_Spiral_primes/src/main.rs
@@ -1,3 +1,4 @@
+#![feature(test)]
 extern crate test;
 extern crate prime;
 
@@ -8,7 +9,7 @@ fn main() {
 		let step_size = len_spir - 1;
 		let first_diag = last_diag - 3*step_size;
 		for num in (first_diag..last_diag).step_by(step_size) {
-			if prime::is_prime(num) { primes_on_diagonals += 1 }
+			if prime::is_prime(num as u64) { primes_on_diagonals += 1 }
 		}
 
 		let amount_diagonal_nums = 2*(len_spir - 1) + 1;
